Enable dynamic switching between English & French
Want to be able to load the string resources for multiple languages simultaneously, compile both files into a single minified file and then select the desired language at runtime.

That's not posible with the existing master implemantion, the last locale file loaded "wins". Need to chang that.

Since right now we only care about English (the default) and French, only modified the main .js file and the French resource file.