Fix edge case bug when picking years
The middle top menu allows you to jump the displayed years ahead by 20 at a time and allows the display of years beyond the max date value.  When it did that the first element of the grid was selected even though it was out of range and pressing enter would put the 'out of range' date into the selector.  Fix that.